Drop redundant qemu-trace-stap copy from qemu-user-static (rhbz#2061584)
The static build of QEMU installs a copy of 'qemu-trace-stap' python
script, which gets renamed to 'qemu-trace-stap-static' by an overly
enthusiastic wildcard. This ends up adding a python dependency to
the qemu-user-static RPM, which is unhelpful.

Anyone who wants to trace QEMU user binaries with the stap helper
can easily install qemu-common as desired.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>