Remove qemu-common as a dep from qemu-user-static
qemu-common has a dep on python, and has nothing that is critical for
the operation of the userspace emulators. At most the qemu-trace-stap
tool is useful, but we shouldn't force install of qemu-common just for
that. qemu-user-static needs to be lightweight as its used to support
cross-arch execution in scenarios where container/image size matters.

In dropping qemu-common as a dep, we just need to ensure we still have
the license files present.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>